Switch canvas from cairo to raqote
This removes the dependency on GTK (pango/cairo), replacing it with an
all-rust rendering based on raqote and ttf-parser.

Important changes because of this:
 - You must specify a list of font files to search instead of font names
 - Font sizes are interpreted differently
 - Markup only supports font= and color= attributes
 - Added text-outline-width to control outline sizes

Fractional point sizes are allowed for fonts and may result in clearer
text than integer sizes.  Font hinting is *not* currently used to
sharpen small characters, so size tweaking is more useful than it
normally is.